Refactor: Custom hook for code submission
