+++ src/app/_components/assignments/student/assignment-layout.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { motion, AnimatePresence } from "framer-motion";
 import { ChevronRight, Maximize2, Minimize2, FileText } from "lucide-react";
 import { Button } from "@/app/_components/ui/button";
@@ -10,9 +10,8 @@ import { QuestionNav } from "./question-nav";
 import { AssignmentById } from "@/lib/types/assignment-tyes";
 import { FullscreenAlert } from "./fullscreen-alert";
 import { CombinedTesting } from "./combined-testing-component";
-import { toast } from "sonner";
 import { useFullScreen } from "@/hooks/use-fullscreen";
-//import { pollJudge0Submissions } from "@/server/actions/submission-actions";
+import { useCodeRunner } from "@/hooks/use-code-runner";
 
 interface AssignmentLayoutProps {
   assignment: AssignmentById;
@@ -26,161 +25,16 @@ export function AssignmentLayout({
   const [currentQuestionIndex, setCurrentQuestionIndex] = useState(0);
   const [isDescriptionExpanded, setIsDescriptionExpanded] = useState(false);
   const [code, setCode] = useState("");
-  const [isRunning, setIsRunning] = useState<boolean>(false);
-  const [testResults, setTestResults] = useState<any[]>([]);
   const [customInput, setCustomInput] = useState("");
-  const [codeStatus, setCodeStatus] = useState<string>("");
   const { isFullscreen } = useFullScreen();
-
   const currentQuestion = assignment.questions[currentQuestionIndex];
-
-  const handleRun = async () => {
-    setIsRunning(true);
-    setCodeStatus("Running");
-
-    try {
-      const response = await fetch("/api/compile", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          code,
-          input: customInput,
-          language: currentQuestion.language,
-        }),
-      });
-      const { output } = await response.json();
-
-      setTestResults([output]);
-    } catch (error: any) {
-      console.error("Error running code:", error);
-      setTestResults([
-        {
-          id: "error",
-          status: "failed" as const,
-          error: error.message || "Failed to run code. Please try again.",
-          runtime: "0s",
-          memory: "0 MB",
-        },
-      ]);
-    } finally {
-      setIsRunning(false);
-    }
-  };
-
-  const handleSubmit = async () => {
-    if (isRunning) {
-      toast.warning("Submission already in progress");
-      return;
-    }
-    setIsRunning(true);
-    setTestResults([]);
-    toast.success("Submitting your solution...");
-    setCodeStatus("Submitting your solution...");
-
-    try {
-      const response = await fetch("/api/submissions", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          code,
-          questionId: currentQuestion.id,
-          language: currentQuestion.language,
-        }),
-      });
-
-      if (!response.ok) {
-        const error = await response.json();
-        throw new Error(error.error || "Failed to submit solution");
-      }
-
-      const data = await response.json();
-      const submissionId = data.submissionId;
-
-      // Start polling for submission status
-      setCodeStatus("Running test cases...");
-      toast.success("Running test cases...");
-      await pollSubmissionStatus(submissionId);
-    } catch (error: any) {
-      console.error("Submission error:", error);
-      setCodeStatus(`Submission failed: ${error.message}`);
-      toast.error(
-        error.message || "An error occurred while submitting your solution",
-      );
-    } finally {
-      setIsRunning(false);
-    }
-  };
-
-  const pollSubmissionStatus = async (submissionId: string) => {
-    let completed = false;
-    let attempts = 0;
-    const maxAttempts = 30; // Poll for maximum of 30 attempts (30 seconds with 2s interval) that is for 60 seconds
-
-    while (!completed && attempts < maxAttempts) {
-      attempts++;
-      await new Promise((resolve) => setTimeout(resolve, 2000)); // Wait for 2 second
-
-      try {
-        const response = await fetch(`/api/submissions/${submissionId}`);
-
-        if (!response.ok) {
-          throw new Error("Failed to fetch submission status");
-        }
-
-        const submissionData = await response.json();
-
-        // Update test results in UI
-        const mappedResults = submissionData.results.map((result: any) => ({
-          id: result.id,
-          status: result.status,
-          runtime: result.runtime,
-          memory: result.memory,
-          output: result.output,
-          error: result.error,
-        }));
-
-        setTestResults(mappedResults);
-
-        // Check if submission is completed
-        if (
-          submissionData.status === "COMPLETED" ||
-          submissionData.status === "PASSED" ||
-          submissionData.status === "FAILED"
-        ) {
-          completed = true;
-
-          // Show appropriate message
-          if (submissionData.status === "COMPLETED") {
-            setCodeStatus("All tests passed successfully!");
-            toast.success("Your solution passed all test cases");
-          } else {
-            setCodeStatus("Some tests failed. Check the results for details.");
-            toast.error("Your solution didn't pass all test cases");
-          }
-
-          break;
-        }
-
-        // await pollJudge0Submissions(submissionId);
-
-        setCodeStatus(`Running test cases (${attempts}/${maxAttempts})...`);
-      } catch (error) {
-        console.error("Error polling submission status:", error);
-      }
-    }
-
-    if (!completed) {
-      setCodeStatus(
-        "Submission is taking longer than expected. You can check results later.",
-      );
-    }
-
-    return completed;
-  };
+  const { isRunning, codeStatus, testResults, runCode, submitCode } =
+    useCodeRunner({
+      code,
+      language: currentQuestion.language,
+      questionId: currentQuestion.id,
+      input: customInput,
+    });
 
   const showFullscreenAlert = assignment.fullScreenEnforcement && !isFullscreen;
 
@@ -269,8 +123,8 @@ export function AssignmentLayout({
                 code={code}
                 onChange={setCode}
                 language={currentQuestion.language}
-                onRun={handleRun}
-                onSubmit={handleSubmit}
+                onRun={runCode}
+                onSubmit={submitCode}
                 isRunning={isRunning}
                 disableCopyPaste={assignment.copyPastePrevention}
               />
@@ -281,7 +135,7 @@ export function AssignmentLayout({
                   results={testResults}
                   customInput={customInput}
                   onCustomInputChange={setCustomInput}
-                  onRunCode={handleRun}
+                  onRunCode={runCode}
                   isRunning={isRunning}
                   codeStatus={codeStatus}
                 />
