+++ src/app/_components/assignments/student/combined-testing-component.tsx
@@ -20,21 +20,11 @@ import {
   AccordionItem,
   AccordionTrigger,
 } from "@/app/_components/ui/accordion";
-
-interface TestResult {
-  id: string;
-  status: "passed" | "failed" | "running";
-  input?: string;
-  runtime?: string;
-  memory?: string;
-  output?: string;
-  error?: string;
-  isCustomInput?: boolean;
-}
+import { CodeRunner } from "@/lib/types/code-types";
 
 interface CombinedTestingProps {
   testCases: TestCase[];
-  results: TestResult[];
+  results: CodeRunner[];
   customInput: string;
   onCustomInputChange: (value: string) => void;
   onRunCode: () => void;
@@ -52,10 +42,8 @@ export function CombinedTesting({
   codeStatus,
 }: CombinedTestingProps) {
   const [isCustomInputOpen, setIsCustomInputOpen] = useState(false);
-
-  // Filter results for test cases vs custom input
-  const customInputResult = results.find((r) => r.isCustomInput);
-  const testCaseResults = results.filter((r) => !r.isCustomInput);
+  const filteredResults = results.filter((result) => !result.hidden);
+  console.log("filteredResults", filteredResults);
 
   return (
     <div className="h-80 overflow-y-auto border-t border-[#2D2D2D] bg-[#1E1E1E]">
@@ -87,7 +75,6 @@ export function CombinedTesting({
         </div>
       </div>
 
-      {/* Custom Input Section */}
       {isCustomInputOpen && (
         <motion.div
           initial={{ opacity: 0, height: 0 }}
@@ -116,87 +103,15 @@ export function CombinedTesting({
                 required by your code.
               </p>
             </div>
-
-            {/* Custom Input Result */}
-            {customInputResult && (
-              <motion.div
-                initial={{ opacity: 0, y: 10 }}
-                animate={{ opacity: 1, y: 0 }}
-                transition={{ duration: 0.2 }}
-                className="mt-2"
-              >
-                <div className="flex items-center gap-3 mb-2">
-                  <div
-                    className={cn(
-                      "flex h-6 w-6 items-center justify-center rounded-full",
-                      customInputResult.status === "passed" && "bg-[#7EBF8E]",
-                      customInputResult.status === "failed" && "bg-[#D2886F]",
-                      customInputResult.status === "running" && "bg-[#F1E6D0]",
-                    )}
-                  >
-                    {customInputResult.status === "passed" && (
-                      <Check className="h-3.5 w-3.5 text-white" />
-                    )}
-                    {customInputResult.status === "failed" && (
-                      <X className="h-3.5 w-3.5 text-white" />
-                    )}
-                    {customInputResult.status === "running" && (
-                      <Clock className="h-3.5 w-3.5 text-[#3A3935]" />
-                    )}
-                  </div>
-                  <div>
-                    <p className="text-sm text-white">Custom Input Result</p>
-                    {customInputResult.status !== "running" &&
-                      customInputResult.runtime &&
-                      customInputResult.memory && (
-                        <p className="text-xs text-[#A1A1A1]">
-                          Runtime: {customInputResult.runtime} | Memory:{" "}
-                          {customInputResult.memory}
-                        </p>
-                      )}
-                  </div>
-                </div>
-
-                {/* Output Section */}
-                {customInputResult.status !== "running" &&
-                  customInputResult.output && (
-                    <div className="mt-3 rounded-md bg-[#2A2A2A] p-3">
-                      <p className="mb-1 text-xs font-medium text-[#A1A1A1]">
-                        Output:
-                      </p>
-                      <pre className="max-h-32 overflow-y-auto whitespace-pre-wrap text-xs text-white font-mono">
-                        {customInputResult.output}
-                      </pre>
-                    </div>
-                  )}
-
-                {/* Error Section */}
-                {customInputResult.status === "failed" &&
-                  customInputResult.error && (
-                    <div className="mt-3 rounded-md bg-[#2A2A2A] border border-[#D2886F]/20 p-3">
-                      <div className="flex items-center gap-2 mb-1">
-                        <AlertCircle className="h-3.5 w-3.5 text-[#D2886F]" />
-                        <p className="text-xs font-medium text-[#D2886F]">
-                          Error:
-                        </p>
-                      </div>
-                      <pre className="max-h-32 overflow-y-auto whitespace-pre-wrap text-xs text-[#D2886F] font-mono">
-                        {customInputResult.error}
-                      </pre>
-                    </div>
-                  )}
-              </motion.div>
-            )}
           </div>
         </motion.div>
       )}
 
-      {/* Test Cases Section */}
       <div className="divide-y divide-[#2D2D2D]">
-        {testCaseResults.length > 0 ? (
-          testCaseResults.map((result, index) => (
+        {filteredResults.length > 0 ? (
+          filteredResults.map((result, index) => (
             <motion.div
-              key={result.id}
+              key={index}
               initial={{ opacity: 0, y: 10 }}
               animate={{ opacity: 1, y: 0 }}
               transition={{ duration: 0.2, delay: index * 0.1 }}
@@ -206,24 +121,24 @@ export function CombinedTesting({
                 <div
                   className={cn(
                     "flex h-6 w-6 items-center justify-center rounded-full",
-                    result.status === "passed" && "bg-[#7EBF8E]",
-                    result.status === "failed" && "bg-[#D2886F]",
-                    result.status === "running" && "bg-[#F1E6D0]",
+                    result?.status === "passed" && "bg-[#7EBF8E]",
+                    result?.status === "failed" && "bg-[#D2886F]",
+                    result?.status === "running" && "bg-[#F1E6D0]",
                   )}
                 >
-                  {result.status === "passed" && (
+                  {result?.status === "passed" && (
                     <Check className="h-3.5 w-3.5 text-white" />
                   )}
-                  {result.status === "failed" && (
+                  {result?.status === "failed" && (
                     <X className="h-3.5 w-3.5 text-white" />
                   )}
-                  {result.status === "running" && (
+                  {result?.status === "running" && (
                     <Clock className="h-3.5 w-3.5 text-[#3A3935]" />
                   )}
                 </div>
                 <div>
                   <p className="text-sm text-white">
-                    {result.status === "running"
+                    {result?.status === "running"
                       ? "Running code..."
                       : `Test Case ${index + 1}`}
                   </p>
@@ -232,9 +147,9 @@ export function CombinedTesting({
                       Hidden
                     </span>
                   )}
-                  {result.status !== "running" &&
-                    result.runtime &&
-                    result.memory && (
+                  {result?.status !== "running" &&
+                    result?.runtime &&
+                    result?.memory && (
                       <p className="text-xs text-[#A1A1A1]">
                         Runtime: {result.runtime} | Memory: {result.memory}
                       </p>
@@ -242,42 +157,42 @@ export function CombinedTesting({
                 </div>
               </div>
 
-              <Accordion type="single" collapsible className="mt-2">
-                <AccordionItem value="input" className="border-[#2D2D2D]">
-                  <AccordionTrigger className="py-2 text-xs text-[#A1A1A1] hover:text-white">
-                    Input
-                  </AccordionTrigger>
-                  <AccordionContent>
-                    <pre className="bg-[#2A2A2A] p-2 rounded text-xs text-white font-mono whitespace-pre-wrap">
-                      {result.input}
-                    </pre>
-                  </AccordionContent>
-                </AccordionItem>
-              </Accordion>
+              {result?.input?.length >= 0 && (
+                <Accordion type="single" collapsible className="mt-2">
+                  <AccordionItem value="input" className="border-[#2D2D2D]">
+                    <AccordionTrigger className="py-2 text-xs text-[#A1A1A1] hover:text-white">
+                      Input
+                    </AccordionTrigger>
+                    <AccordionContent>
+                      <pre className="bg-[#2A2A2A] p-2 rounded text-xs text-white font-mono whitespace-pre-wrap">
+                        {result?.input}
+                      </pre>
+                    </AccordionContent>
+                  </AccordionItem>
+                </Accordion>
+              )}
 
               {/* Output Section */}
-              {result.status !== "running" &&
-                result.output &&
-                !testCases[index]?.hidden && (
-                  <div className="mt-3 rounded-md bg-[#2A2A2A] p-3">
-                    <p className="mb-1 text-xs font-medium text-[#A1A1A1]">
-                      Output:
-                    </p>
-                    <pre className="max-h-32 overflow-y-auto whitespace-pre-wrap text-xs text-white font-mono">
-                      {result.output}
-                    </pre>
-                  </div>
-                )}
+              {result?.status !== "running" && result?.output && (
+                <div className="mt-3 rounded-md bg-[#2A2A2A] p-3">
+                  <p className="mb-1 text-xs font-medium text-[#A1A1A1]">
+                    Output:
+                  </p>
+                  <pre className="max-h-32 overflow-y-auto whitespace-pre-wrap text-xs text-white font-mono">
+                    {result.output}
+                  </pre>
+                </div>
+              )}
 
               {/* Error Section */}
-              {result.status === "failed" && result.error && (
+              {result?.status === "failed" && result?.error && (
                 <div className="mt-3 rounded-md bg-[#2A2A2A] border border-[#D2886F]/20 p-3">
                   <div className="flex items-center gap-2 mb-1">
                     <AlertCircle className="h-3.5 w-3.5 text-[#D2886F]" />
                     <p className="text-xs font-medium text-[#D2886F]">Error:</p>
                   </div>
                   <pre className="max-h-32 overflow-y-auto whitespace-pre-wrap text-xs text-[#D2886F] font-mono">
-                    {result.error}
+                    {result?.error}
                   </pre>
                 </div>
               )}
