+++ src/app/_components/ui/sonner.tsx
@@ -10,6 +10,7 @@ const Toaster = ({ ...props }: ToasterProps) => {
 
   return (
     <Sonner
+      richColors
       theme={theme as ToasterProps["theme"]}
       className="toaster group"
       toastOptions={{
