+++ src/lib/types/assignment-tyes.ts
@@ -25,28 +25,6 @@ export interface Assignment {
   fullScreenEnforcement: boolean;
 }
 
-export interface TestResults {
-  description: string;
-  passed: boolean;
-  isBonus: boolean;
-  executionTime: number | null;
-  error: string | null;
-}
-
 export interface AssignmentById extends Assignment {
   questions: Question[];
-}
-
-export interface judgeResult {
-  stdout: string;
-  time: string;
-  memory: number;
-  stderr: string | null;
-  token: string;
-  compile_output: string | null;
-  message: string | null;
-  status: {
-    id: number;
-    description: string;
-  };
 }
